refactored account components to create register/login from one
